third_party/WebKit/LayoutTests/http/tests/security/link-crossorigin-stylesheet-use-credentials.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
<script>
description("Test that a linked stylesheet with a crossorigin=use-credentials attributes loads expected CORS enabled resources.")
</script>
<script src="resources/link-crossorigin-common.js"></script>
</head>
<body>
<link crossorigin="use-credentials" rel="stylesheet" href="http://localhost:8080/security/resources/green-background-allow-credentials.php" onload="pass()" onerror="fail()">
<link crossorigin="use-credentials" rel="stylesheet" href="http://localhost:8080/security/resources/green-background-allow-star.php" onload="fail()" onerror="pass()">
<script>
function runTest()
{
    // Test that dynamically inserted <link> elements are handled the same way.
    var link = document.createElement("link");
    link.rel = "stylesheet";
    link.crossOrigin = "use-credentials";
    link.onload = pass;
    link.onerror = fail;
    link.href = "http://localhost:8080/security/resources/green-background-allow-credentials.php";
    document.body.appendChild(link);

    link = document.createElement("link");
    link.rel = "stylesheet";
    link.crossOrigin = "use-credentials";
    link.onload = fail;
    link.onerror = pass;
    link.href = "http://localhost:8080/security/resources/green-background-allow-star.php";
    document.body.appendChild(link);
}
window.onload = runTest;
</script>
</body>
</html>
